[PA-21] Use message based prediction
