+++ tf_kit/model.py
@@ -18,7 +18,7 @@ def get_lstm():
     ssm.add(tf.keras.layers.Dense(32))
     ssm.add(tf.keras.layers.Dense(16))
     ssm.add(tf.keras.layers.Dense(N_CLASSES, activation='softmax'))
-    ssm.compile(optimizer=tf.keras.optimizers.Adam(lr=0.003),
+    ssm.compile(optimizer=tf.keras.optimizers.Adam(lr=0.001),
                 loss=tf.keras.losses.CategoricalCrossentropy(),
                 metrics=['accuracy', tfa.metrics.F1Score(num_classes=N_CLASSES)])
     return ssm
